refactor: Use verification code generator as util
