+++ src/main/java/com/marceldev/companylunchcomment/util/VerificationCodeGenerator.java
@@ -1,12 +1,10 @@
-package com.marceldev.companylunchcomment.component;
+package com.marceldev.companylunchcomment.util;
 
 import java.util.Random;
-import org.springframework.stereotype.Component;
 
-@Component
 public class VerificationCodeGenerator {
 
-  public String generate(int length) {
+  public static String generate(int length) {
     StringBuilder code = new StringBuilder();
     new Random().ints(length, 0, 10)
         .forEach(code::append);
